CSVハンドラ系リファクタ
